Closed Issue #8:
The *tethys gen apache* script would throw an error when executed as directed in the production docs:

::

    tethys gen apache -d /etc/apache2/site-available/tethys-default.conf
    ERROR: "{0}" is not a valid directory.

The instructions in the docs were bad, because the *-d* option takes the path to a directory, not the file to create. This line of the docs was updated as follows:

::

    tethys gen apache -d /etc/apache2/sites-available

The error message made this issue unclear, because the directory was not being injected into the error message. This has been resovled so that the error now would read:

::

    ERROR: "/etc/apache2/sites-available/tethys-default.conf" is not a valid directory.