Database model for apps implemented including refactoring methods that use SingletonAppHarvester to get a list of apps to use DB lookup instead. However, could not refactor cases that use the SingletonAppHarvester to call methods on the App class ( e.g. app.url_maps() ).
One possible solution is to store results of those method calls in db and write shadow methods on the TethysApp model that will pull them from the db rather than the singleton... needs to be decided on a case by case basis. For example the url_maps are only used at loading time, so we don't need to store in db.