Tethys Skin Update (#265)
Updated app skin for Tethys 2.0. Includes leaner header and app actions section, redesigned nav toggle and exit buttons, a new button in the header for quick access to settings, and a new block for adding custom buttons to the header. Pertinent documentation has been updated as well.

In addition the scaffold has been revamped to include demonstrations of proper usage of modals and the new header_buttons block. The app_base.css is minified and the header and app_actions section have been refactored to use flex box css styling. The random colors that new apps are generated with have been changed as well.

This pull request also includes commits that address #127, #258, and #246.