Security Updates (#542)
* Fix bug with scaffolding extensions. (#534)

* Increment patch for hotfix 3.0.1

* Added autocomplete: off attribute to all password fields.
Most browsers have a facility to remember user credentials that are entered into HTML forms. This function can be configured by the user and also by applications that employ user credentials. If the function is enabled, then credentials entered by the user are stored on their local computer and retrieved by the browser on future visits to the same application.
The stored credentials can be captured by an attacker who gains control over the user's computer. Further, an attacker who finds a separate application vulnerability such as cross-site scripting may be able to exploit this to retrieve a user's browser-stored credentials.

* Added @no_cache decorator to views involving login credentials to prevent browsers and other cache systems from storing user credentials that can potentially be retrieved by another individual.

* Finished adding no_cache cache control header to views with sensitive information.

* Modified permission_required decorator to return Error 404: Not Found and similar message when DEBUG is False to address Directory Enumeration vulnerability.

* Fixed @never_cache decorator usage bug

* Altered message to just reflect a 404 Not Found message as suggested by @sdc50

* revert to 3.1.0

* Security Fixes (#543)

* Fix bug with scaffolding extensions. (#534)

* Increment patch for hotfix 3.0.1

* Added autocomplete: off attribute to all password fields.
Most browsers have a facility to remember user credentials that are entered into HTML forms. This function can be configured by the user and also by applications that employ user credentials. If the function is enabled, then credentials entered by the user are stored on their local computer and retrieved by the browser on future visits to the same application.
The stored credentials can be captured by an attacker who gains control over the user's computer. Further, an attacker who finds a separate application vulnerability such as cross-site scripting may be able to exploit this to retrieve a user's browser-stored credentials.

* Added @no_cache decorator to views involving login credentials to prevent browsers and other cache systems from storing user credentials that can potentially be retrieved by another individual.

* Finished adding no_cache cache control header to views with sensitive information.

* Modified permission_required decorator to return Error 404: Not Found and similar message when DEBUG is False to address Directory Enumeration vulnerability.

* Fixed @never_cache decorator usage bug

* Altered message to just reflect a 404 Not Found message as suggested by @sdc50

* revert to 3.1.0

Co-authored-by: Nathan Swain <swainn@users.noreply.github.com>

* Various Fixes for the Documentation (#532)

* Update Django docs links.

* Fixes various typos.
Adds shortcut link to app migration documentation.

* Adds links to presentations for Google Earth Engine and THREDDS tutorials
Adds links ot presentations with overviews of Tethys 3 and app migration

* Make instructions for UrlMaps more clear in THREDDS and GEE tutorials.

* Fixes to the Tethys Portal Configuration settings documentation.

* Make JavaScript instructions more clear in the GEE tutorial.

* Make JavaScript instructions more clear in the THREDDS tutorial.

* Add data mocking tip to THREDDS tutorial.

* Revert "Security Fixes (#543)"

Tests need to be updated and coverage back to 100%.

This reverts commit f7734a3a

* Fixed failing unit tests.

* use Bokeh AutoloadJsConsumer with cdn (#546)

Co-authored-by: msouff <msouffront@aquaveo.com>

* Added autocomplete: off attribute to all password fields.
Most browsers have a facility to remember user credentials that are entered into HTML forms. This function can be configured by the user and also by applications that employ user credentials. If the function is enabled, then credentials entered by the user are stored on their local computer and retrieved by the browser on future visits to the same application.
The stored credentials can be captured by an attacker who gains control over the user's computer. Further, an attacker who finds a separate application vulnerability such as cross-site scripting may be able to exploit this to retrieve a user's browser-stored credentials.

* Added @no_cache decorator to views involving login credentials to prevent browsers and other cache systems from storing user credentials that can potentially be retrieved by another individual.

* Finished adding no_cache cache control header to views with sensitive information.

* Modified permission_required decorator to return Error 404: Not Found and similar message when DEBUG is False to address Directory Enumeration vulnerability.

* Fixed @never_cache decorator usage bug

* Altered message to just reflect a 404 Not Found message as suggested by @sdc50

* Fixed failing unit tests.

Co-authored-by: Nathan Swain <swainn@users.noreply.github.com>
Co-authored-by: nswain <nswain@aquaveo.com>
Co-authored-by: Michael Souffront <masouffront@gmail.com>
Co-authored-by: msouff <msouffront@aquaveo.com>