Broken URIs (#547)
* Fix bug with scaffolding extensions. (#534)

* Increment patch for hotfix 3.0.1

* Added autocomplete: off attribute to all password fields.
Most browsers have a facility to remember user credentials that are entered into HTML forms. This function can be configured by the user and also by applications that employ user credentials. If the function is enabled, then credentials entered by the user are stored on their local computer and retrieved by the browser on future visits to the same application.
The stored credentials can be captured by an attacker who gains control over the user's computer. Further, an attacker who finds a separate application vulnerability such as cross-site scripting may be able to exploit this to retrieve a user's browser-stored credentials.

* Added @no_cache decorator to views involving login credentials to prevent browsers and other cache systems from storing user credentials that can potentially be retrieved by another individual.

* Finished adding no_cache cache control header to views with sensitive information.

* Modified permission_required decorator to return Error 404: Not Found and similar message when DEBUG is False to address Directory Enumeration vulnerability.

* Fixed @never_cache decorator usage bug

* Altered message to just reflect a 404 Not Found message as suggested by @sdc50

* revert to 3.1.0

* Fixed failing unit tests.

* Added autocomplete: off attribute to all password fields.
Most browsers have a facility to remember user credentials that are entered into HTML forms. This function can be configured by the user and also by applications that employ user credentials. If the function is enabled, then credentials entered by the user are stored on their local computer and retrieved by the browser on future visits to the same application.
The stored credentials can be captured by an attacker who gains control over the user's computer. Further, an attacker who finds a separate application vulnerability such as cross-site scripting may be able to exploit this to retrieve a user's browser-stored credentials.

* Added @no_cache decorator to views involving login credentials to prevent browsers and other cache systems from storing user credentials that can potentially be retrieved by another individual.

* Finished adding no_cache cache control header to views with sensitive information.

* Modified permission_required decorator to return Error 404: Not Found and similar message when DEBUG is False to address Directory Enumeration vulnerability.

* Fixed @never_cache decorator usage bug

* Altered message to just reflect a 404 Not Found message as suggested by @sdc50

* Fixed failing unit tests.

* Added tests and fix for #545 - Password Reset URLs without pattern name.

* Split test_permission_required_exception test into two tests for production and dev settings coverage.

Co-authored-by: Nathan Swain <swainn@users.noreply.github.com>