Fix error message showing in green box (#908)
Before this change, an exception during delete would show the error message in a green success box and remove the job row from the browser display, although a browser refresh would show that job again. After this change, it will show the error in a red box and still display the job.